Create README.md
# finalProject
Initially when we started our work on the project we completed first steps which are works on the logic of the program. So we created tables in database and connection among them
Also, we looked on the logic of whole program and functionality based on the requirements provided by teacher.
Next step is work on entitites.
A bit further we almost finished controllers,validator, repositories, interfaces
The last step to finish our work is application which is the key points of the work which was difficult one.
As the final stage we checked our code together on the existence of errors. In addition, we made our code a bit more readable, fix wrong structures.
Our project is about bank management system. Project is connected with DataBase which consist of tables users,cards,transactions and debts. Our program allows users to create an account in the system if they do not have and log in in created or existing account.
Furthemorer our application provides an opportunity to check all the cards of particular use or create card(s) if user does not have any card.
Program allows to check available balance on account, check loans, take a new loan, calculate loan plan based on the time for which user takes a loan.
Also user can make a different type of transactions and view history of completed transactions if user had any including transaction information such as date,time,type and amount.